Improve and fix natural mergesort for linked lists
In the previous commit, I improved the natural mergesort
implementation for arrays, and verified that it achieved
linearithmic performance via a doubling test. The data
supported the linearithmic hypothesis.

In this commit, I have applied my learnings from the array
version, and similarly added a doubling test that supports
the hypothesis that the performance is linearithmic. In the
process I also identified an error in my implementation that
led to incorrect results because some nodes in the list
were essentially lost due to the mishandling of the "next"
references.